Medical_AI/requirements.txt
gradio
pandas
google-generativeai